modul_3/veb3_api/requirements.txt
python-dotenv
requests
sdamgia-api
pytesseract